refactored relevant_recipe into its own func
