+++ streamlit_app/sql_queries.py
@@ -1,9 +1,17 @@
-from typing import List
+from typing import List, Tuple
+from dataclasses import dataclass
 
-def top_ingredients_query(include_unspecified_meal: bool, include_unspecified_diet: bool) -> str:
+@dataclass
+class SearchTags:
+    diets: Tuple[str] = tuple()
+    include_unspecified_diets: bool = False
+    meals: Tuple[str] = tuple()
+    include_unspecified_meals: bool = False
+
+def relevant_recipes(include_unspecified_meal: bool, include_unspecified_diet: bool) -> str:
     extra_meal_str = "OR meal.name IS NULL" if include_unspecified_meal else ""
     extra_diet_str = "OR diet.name IS NULL" if include_unspecified_diet else ""
-    query = f"""
+    relevant_recipes_query = f"""
     WITH relevant_recipe AS (
         SELECT recipe.id FROM recipe
         LEFT JOIN meal ON meal.recipe_id = recipe.id
@@ -12,8 +20,14 @@ def top_ingredients_query(include_unspecified_meal: bool, include_unspecified_di
             (meal.name IN %s {extra_meal_str}) AND
             (diet.name IN %s {extra_diet_str})
         GROUP BY recipe.id
-    )
+    )"""
+    return relevant_recipes_query
 
+
+def top_ingredients_query(include_unspecified_meal: bool, include_unspecified_diet: bool) -> str:
+    relevant_recipes_definition = relevant_recipes(include_unspecified_meal, include_unspecified_diet)
+    query = f"""
+    {relevant_recipes_definition}
     SELECT
         ingredient.name,
         COUNT(relevant_recipe.id)
@@ -26,18 +40,9 @@ def top_ingredients_query(include_unspecified_meal: bool, include_unspecified_di
     return query
 
 def top_equipment_query(include_unspecified_meal: bool, include_unspecified_diet: bool) -> str:
-    extra_meal_str = "OR meal.name IS NULL" if include_unspecified_meal else ""
-    extra_diet_str = "OR diet.name IS NULL" if include_unspecified_diet else ""
+    relevant_recipes_definition = relevant_recipes(include_unspecified_meal, include_unspecified_diet)
     query = f"""
-    WITH relevant_recipe AS (
-        SELECT recipe.id FROM recipe
-        LEFT JOIN meal ON meal.recipe_id = recipe.id
-        LEFT JOIN diet ON diet.id = recipe.diet_id
-        WHERE
-            (meal.name IN %s {extra_meal_str}) AND
-            (diet.name IN %s {extra_diet_str})
-        GROUP BY recipe.id
-    )
+    {relevant_recipes_definition}
 
     SELECT
         equipment.name,
@@ -52,43 +57,32 @@ def top_equipment_query(include_unspecified_meal: bool, include_unspecified_diet
     return query
 
 def num_recipes_selected(include_unspecified_meal: bool, include_unspecified_diet: bool) -> str:
-    extra_meal_str = "OR meal.name IS NULL" if include_unspecified_meal else ""
-    extra_diet_str = "OR diet.name IS NULL" if include_unspecified_diet else ""
+    relevant_recipes_definition = relevant_recipes(include_unspecified_meal, include_unspecified_diet)
     query = f"""
-    SELECT COUNT(DISTINCT(recipe.id)) FROM recipe
-    LEFT JOIN meal ON meal.recipe_id = recipe.id
-    LEFT JOIN diet ON diet.id = recipe.diet_id
-    WHERE
-        (meal.name IN %s {extra_meal_str}) AND
-        (diet.name IN %s {extra_diet_str})
+    {relevant_recipes_definition}
+
+    SELECT COUNT(DISTINCT(id)) FROM relevant_recipe;
     """
     return query
 
 
 def times(include_unspecified_meal: bool, include_unspecified_diet: bool) -> str:
-    extra_meal_str = "OR meal.name IS NULL" if include_unspecified_meal else ""
-    extra_diet_str = "OR diet.name IS NULL" if include_unspecified_diet else ""
+    relevant_recipes_definition = relevant_recipes(include_unspecified_meal, include_unspecified_diet)
     query = f"""
+    {relevant_recipes_definition}
+
     SELECT prep_time, cook_time FROM recipe
-    LEFT JOIN meal ON meal.recipe_id = recipe.id
-    LEFT JOIN diet ON diet.id = recipe.diet_id
-    WHERE
-        (meal.name IN %s {extra_meal_str}) AND
-        (diet.name IN %s {extra_diet_str})
-    GROUP BY recipe.id;
+    INNER JOIN relevant_recipe ON recipe.id = relevant_recipe.id;
     """
+    print(query)
     return query
 
 def ingredient_search_query(num_terms: int) -> str:
-
     sub_query = "SELECT recipe_id FROM ingredient WHERE name LIKE %s"
-
     where_clause = f"WHERE id IN ({sub_query})" if num_terms > 0 else ""
-
     if num_terms > 1:
         for i in range(num_terms-1):
             where_clause = where_clause + f" AND id IN ({sub_query})"
-
     query = f"""
         SELECT title, url FROM recipe
         {where_clause}
